New backend that allows UDFs to be written in C/C++
The new backend compiles the user-defined-function into a shared
library object using the g++ compiler. That object gets written
to the dataset payload area, just like we do with LuaJIT's bytecode.

When the filter is invoked to read a UDF dataset we dlopen() and
retrieve the needed symbols using dlsym(), and then invoke the
UDF entry point.

The template file for C/C++ has useful routines that let the user
retrieve datasets parsed by the HDF5-UDF engine as part of the
dependency resolution, and also to retrieve dimensions by dataset
name.