Add support for user-defined-functions in Python.
This change adds initial support for writing UDF datasets in Python.
Just like we do in the other backends, we merge the UDF template file
with the user-provided code and compile them into a Python bytecode
(.pyc) using the 'compileall' module. The resulting bytecode is
embedded in the HDF5 file, and retrieved on-the-fly when the
application requests to read the corresponding dataset.

This commit does not verify the Python bytecode version. It is assumed
that compatible versions of the CPython interpreter are used on the
machines that create the UDF datasets and that read it. This
implementation also expects the code object to start at offset 16 of the
bytecode file. Different versions of Python may have a bytecode header
of different size; this assumption needs to be revisited at some point.