Lua: reuse methods from Backend + namespace changes
This commit changes the Lua code so it uses the new
Backend::assembleUDF method. This allows the backends to reuse functions
that are common to them.

The function names exported by the library are now prefixed by 'lua'
to prevent clashes with their Python counterparts: get_data() became
luaGetData(), get_type() became luaGetType(), and so on.