Initial support for syscall filtering.
This commit adds support for system call filtering to the C/C++ backend
through seccomp and the syscall_intercept library. We define the list of
syscalls that UDFs are allowed to invoke (sandbox.cpp) and inspect the
path names provided to potentially dangerous calls such as open
(syscall_wrappers.cpp).

The UDF now executes as a separate process. If seccomp finds that the
UDF is attempting to execute a non-allowed syscall that process is
terminated without crashing the main program.

Now that the foundation is laid out we can look into making the
implementation more customizable by (1) using a config file to tell
which syscalls are OK to be called by UDFs, (2) determining at compile
time whether to use seccomp or not, and (3) extending support for
syscall filtering to the Lua and Python backends.

Refs #2