Makefile: enable conditional build of syscall filtering
It is now possible to run `make OPT_SANDBOX=1` to build HDF5-UDF with
support for system call filtering through seccomp + syscall_intercept.
Running `make OPT_SANDBOX=0` produces a build that will neither install
seccomp rules at runtime, nor link UDFs to libsyscall_intercept.

Note that there is still room for improvement: if OPT_SANDBOX=1 then we
should be able to check that the UDF is really linked against
libsyscall_intercept. If that is not the case, then we should either
prevent the UDF from running or a more strict set of syscall rules
should be put in place to run that UDF.

Refs #2