Pack syscall-intercept/seccomp library into libhdf5-udf
Because syscall-intercept patches the image of the process at the very
instant the library is loaded into memory, we have to be careful about
the way in which we open it. Prior to this commit we were compiling the
UDF code and syscall-intercept together, and embedding the result into
the HDF5 dataset.

There are two problems with that approach. First, the size of the
dataset becomes larger. If we have two or more UDF datasets then we will
end up with two or more copies of the syscall-intercept library in the
HDF5 file. Second, we cannot really tell if a HDF5 file provided by a
third-party has an UDF that was linked to syscall-intercept as we expect.

The new approach is to compile a stand-alone library that initializes
syscall-intercept and that contains the seccomp rules. That library is
then appended to libhdf5-udf.so (our HDF5 filter) under a special
section of the ELF file named '.hdf5-udf-sandbox'. At run-time, when the
UDF code is getting prepared to be executed, we extract the data from
that section and dlopen() it in order to initialize the seccomp rules
and place the hooks into the syscalls issued by the UDF process.

This change comes with two bonuses: (1) it becomes easier to provide
sandboxing support for the other backends, such as Python and Lua, and
(2) distribution of the software is much easier, as we have a single
shared library to install.

Refs #2