Add a sandbox the Python backend.
This code change protects the execution of Python UDFs by employing the
same sandbox infrastructure that we use for C/C++ UDFs. Note that we had
to add one extra system call to the allow-list (exitgroup()) and that we
are assuming that the udf_template.py compiled with the UDF is
trustable. It is possible that a malicious user had fiddled with the
udf_template that's shipped with the UDF, though, so we need to revisit
this in the near future.

The reason why we don't execute the functions from the udf_template in a
sandbox as well is because they import Python modules, and that requires
access to several files in the filesystem. We would end up with a very
permissive sandbox if we were to follow that path.

Refs #2