Add missing file that implements AnonymousMemoryMap.
The implementation of the helper class AnonymousMemoryMap used to live
in the C/C++ backend, but now that it's shared with the Python backend
it makes sense to move it to its own file.

Note that the sole reason for having this simple class is to let its
destructor take care of munmap-ing previously mmap-ed areas. This makes
it much easier to deal with error paths..

Refs #2