Add a sandbox to the LuaJIT backend.
As with the previous commits, this one protects the calls to the Lua UDF
with seccomp and syscall-intercept. Even though we use LuaJIT and it
comes with a virtual machine that provides isolation, it is well known
that it is possible to escape its sandbox. We expect with this code
change that overall execution of UDFs get more robust and secure for
those opening UDF-based datasets provided by third-party.

Refs #2.