Enable conditional compilation of backends.
It is now possible to disable the compilation of each backend with the
following arguments to make: OPT_PYTHON=0, OPT_LUA=0, OPT_CPP=0. By
default all backends will be built.

The documentation has been changed accordingly. Build instructions,
which used to be in README.md, are now in REQUIREMENTS.md (to be renamed
to INSTALL.md soon).