Fixes warnings with the Lua backend.
The Lua backend appends an empty DatasetInfo entry to the Lua table as
an indicator of the last valid member. However, when calling the
getDatatype() method of DatasetInfo on an empty entry, we will issue a
call to H5Tequal(-1, id), which results in a warning coming from
libhdf5.

This commit fixes that warning by properly validating the data on each
get() method of that class.