Add support for scratch datasets.
It is possible to have two or more output datasets in a UDF script, as
in the following trivial example:

```
function dynamic_dataset()
    local out1 = lib.getData("DynamicDataset1")
    local out2 = lib.getData("DynamicDataset2")

    for i = 0, lib.getSize("DynamicDataset1")[1] do
        out1[i] = i
        out2[i] = out1[i]*2
    end
```

When processed by `hdf5-udf`, two new datasets will be appended to the
HDF5 file while having the very same bytecode. However, when the
application asks to read `DynamicDataset1` we cannot retrieve
`DynamicDataset2`, as that will lead to an infinite loop. Rather, we
want to allocate a temporary buffer for `DynamicDataset2` and discard it
once the function returns.

This commit does exactly that by creating "scratch datasets". A new
entry in the JSON metadata indicates which datasets have to be allocated
that way so that the UDF handler knows how to prepare the data for the
user function.