Improve build on other distros.
Starting with Python 3.8, `python3-config --ldflags` no longer links
to libpython; one must now use `python3-config --embed` in addition
to `--ldflags` to compile C extensions. This commit tests if the
`python3-config` checks the new flag and uses it if it does.

This commit also replaces the explicit link to libhdf5.so with a call
to `pkg-config` so that the right library name is picked. The same
applies to CFLAGS so that packages installed on alternative locations
can be found.

Thanks to @snhardin for raising the build problem on Debian/Ubuntu
based distros in first place.