Fallback to /proc/self/exe if realpath(argv[0]) fails
We use realpath() to identify the directory where the hdf5-udf file
was installed so we can tell where the UDF template files were put.
If the program name does not contain any elements to be resolved,
then realpath() does not do anything, though -- and we can't tell
where the template files are.

This commit uses /proc/self/exe as a fallback in case realpath on
argv[0] does not return the information we need.