Implement a RPM spec file for the project
This commit introduces a spec file for the creation of RPM packages
for HDF5-UDF. The spec file depends on the syscall_intercept, for
which no official RPM package exists (yet). Thus, we include the spec
file for syscall_intercept, too. A Dockerfile is included to automate
building those packages.

The hdf5.pc file included in the SOURCES directory is required during
the compilation of HDF5-UDF -- that file is not shipped by the HDF5
package in Fedora, so we have to provide our own.

Note that the RPM and the Dockerfile currently download a pre-release
tarball from our page on GitHub. The URL will be changed once the final
release is tagged.

Closes #4