Fixes issue with --overwrite.
When a dataset name is provided in the command line and that dataset
already exists on the HDF5 file, then do not treat it as a potential
input dataset: it's gonna be an output dataset for sure.

The problem of treating output datasets as input is that we keep some
handles of input datasets opened, and that leads to datasets with an
invalid state after H5Ldelete() is called.

Closes #6