Seccomp: send SIGSYS instead of killing the UDF.
The disadvantage of killing the process upon the execution of a system
call that's not allowed to run is that the user never gets to know what
operation caused the UDF to fail. Thankfully, Seccomp allows a signal to
be delivered to the process instead; the siginfo structure provided to
the signal handler includes the system call number that the process
tried to run, which we use to give a hint to the UDF programmer. The UDF
process exits after the error message is shown in the console.

Another change introduced by this commit relates to the allow-list. If
a syscall attempts to access a file that's not allowed to be read, then
syscall_intercept returns -EACCESS instead of -EPERM.