Python: run lib.load() protected by the Sandbox.
This commit ensures that both the `dynamic_dataset()` callback and
`lib.load()` are called from a sandboxed process.

To do so, we have to move away some code from the Python template
(`udf_template.py`) and reimplement them in C -- otherwise the sandbox
that executes the template and the UDF requires access to way too many
system calls and filesystem objects. The following Python calls are now
handled from the CPython API:

1. `import cffi`
2. `self.ffi = cffi.FFI()`
3. `self.filterlib = self.ffi.dlopen(path)`