Fedora: set PLUGINDIR at install time, disable strip
Stripping of libraries breaks libhdf5-udf.so, because we store the
sandbox library as a special section of the ELF file. This commits
disables building the debuginfo package, as that seems to be the only
way to prevent strip from being called on the binaries and libraries.

We also set PLUGINDIR at install time so that the HDF5 I/O filter is
installed at the default location searched by the HDF5 Fedora package
(i.e., under /usr/local/hdf5/lib/plugin).