Python: search for other names of libpython.so
Some distros provide a libpython3.so symlink that points to the actual
libpython${major}.${minor}.so file. Other distros, such as Ubuntu, do
not. To prevent dlopen() from failing, we now attempt to load both
libpython3.so and libpython3.${minor}. This fixes the problem of
HDF5-UDF not finding symbols from libpython when loading the cffi
module.