Python: add support for compound data structures
This commit extends support for compound types by letting the Python
backend recognize them. The UDF looks exactly the same as C/C++: a call
to getData() returns a data structure, arranged by CFFI, that gives
access to each member of the original compound:

```
def dynamic_dataset():
    compound = lib.getData("Dataset1")
    udf_data = lib.getData("Temperature")
    udf_dims = lib.getDims("Temperature")

    for i in range(udf_dims[0]):
        udf_data[i] = compound[i].temperature
```

The structure declaration is injected in a new placeholder on
`udf_template.py`. Some minor changes were made to the Lua and C/C++
backends to take into account the existence of that placeholder.