Lua: add support for compound data structures
This commit extends support for compound types by letting the Lua
backend recognize them. The UDF looks exactly the same as C/C++ and
Python: a call to getData() returns a data structure, arranged by
FFI, that gives access to each member of the original compound:

```
function dynamic_dataset()
    local compound = lib.getData("Dataset1")
    local udf_data = lib.getData("Temperature")
    local udf_dims = lib.getDims("Temperature")

    for i=0, udf_dims[1]-1 do
        udf_data[i] = compound[i].temperature
    end
end
```

The structure declaration is injected in a new placeholder on
`udf_template.lua`.