getProcCandidates: disable error messages when probing files
Because the I/O filter API does not provide the file handle to the
filter, we scan /proc/self/fd/ for file descriptors opened by the
application and then attempt to open the corresponding file. If HDF5
successfully opens that file and the dataset we're opening does exist on
that file, then we know we found it.

However, when the program has other non-HDF5 files opened, then the call
to H5Fopen() will naturally fail -- and the users will see plenty of
error messages that may confuse them. This commit temporarily silents
error reporting while we call H5Fopen() on files from /proc/self/fd/. We
also fix a call to H5Fclose() so that it's only performed when we have a
valid handle to close.