Python: fix reference counting
Decrement reference count of three objects: the bytecode object, the
'dlopen' attribute borrowed from the cffi.FFI() instance, and the
'load' attribute from the UDF template code. We also stop decrementing
the reference count of objects passed to PyTuple_SetItem(), as that
function steals the reference to the original object. Attempting to
decrement it once again may lead to invalid memory access and potential
crashes when Py_Finalize() runs.