Compound/C++: add support for strings
This commit allows HDF5-UDF to recognize strings in compounds. Both
variable-sized and fixed-sized strings are handled by this
implementation.

The C structure created from the compound needs to be aligned at a byte
boundary so that we can properly iterate over the elements read by
H5Dread(). We use GCC's __attribute__((packed)) for that purpose. When
creating that structure we also declare either a char array with a
predetermined size or a char*, depending on the HDF5 datatype
underneath.