Compound/Lua: add support for strings
This commit extends support for strings in compounds by implementing it
in the Lua backend. Both variable-sized and fixed-sized strings are
handled by this implementation.

The C structure created from the compound needs to be aligned at a byte
boundary so that we can properly iterate over the elements read by
H5Dread(). We use '#pragma pack(1)' for that purpose, as that's the
mechanism supported by LuaJIT's FFI cdef() parser.