C++: strip symbol table and relocation info from UDF
Size matters: we want to keep the UDF shared library as small as we can.
We already compile with -Os and compress the output library prior to
saving it to disk, but somehow we missed to include '-s' on our original
optimization list. Doing so reduces the output by stripping its symbol
table and relocation information. Some of the example UDFs have been
shrinked by ~25% with this code change.