Tests: new script to check for regressions
The new `run_tests.sh` script runs a few different UDFs and compares
their output against a reference set. The script allows the comparison
of the output of `h5dump` as well as of `stdout` (in which case the UDF
outputs elements read with printf). All backends are supported by this
simple test framework.