Tests: keep original HDF5 file names
For simplicity, we were renaming the example files build from the
examples/ diretory so that the 'example-' prefix was removed. We were
using CoreUtils' `rename` utility for that task, but as it turns out
some distros ship a Perl module named `rename` as well -- which has
different semantics than the one from CoreUtils. This commit simply
keeps the original file names to avoid running into silly problems
across Linux distros.