Add support for string datasets outside of compounds
It is now possible to use fixed- and variable-sized string datasets as
input to UDFs. Because the way they are declared is different (`char *`
versus `char name[len]`), we embed the string in a packed structure so
that the UDF can easily iterate over the input strings using simple
pointer arithmetic.

The generated structure is named after the sanitized dataset name
followed by '_t'. The string data is then accessible through that
structure's `value` member. For instance, a dataset named `Foo` declared
as a string with a fixed size of 32 bytes would be converted into the
following structure by HDF5-UDF:

```
struct foo_t {
    char value[32];
};
```

Note that, depending on the backend, explicit struct packing attributes
are included in the declaration of the generated structure.