Lua: support taking input from string datatypes
This commit adds support for reading string-based datasets on UDFs. The
API is the same one used by Python: `lib.getData("datasetName")` gives
the user a pointer to an array of structures that embeds the input
strings; `lib.string()` can then be called on different indexes of that
array to convert the given member to a string object.

Accompanying tests are also included in this commit to check for any
potential regressions in the future.