Tests: use numdiff to account for float precision errors
We now check the output of the test UDFs with numdiff. That tool is able
to detect numerical fields on the input files and to compute the
difference between numbers across the files while honoring a
configurable error tolerance threshold. If we still detect a difference
between two files then we resort to `diff -up` to show where the
differences are.