Add support for outputting string datatypes
It is now possible to write UDFs that output strings on all backends.
The same infrastructure used to read string-based datasets is used:
strings are wrapped around a named structure with a single member named
'char value[]'.

For this initial implementation, the size of the string is predefined
at 32 bytes. It would be great if we could use variable-sized strings,
but the problem is that the call to H5Dwrite() needs to learn about the
size of each string member of the data buffer, and it does so with calls
to strlen(). Because we place the bytecode data on that buffer, such
calls are likely to fail and crash the program, hence a static string
size needs to be used. A future code change will allow the UDF writer to
specify the size of the string.

Another limitation for now is that we don't have any wrappers such as
lib.string() to hide the structure members and ease writing the UDF. For
instance, instead of:

    snprintf((char *) data[i].value, sizeof(data[i].value), "%s", str)

we would like to provide a helper function such as:

    lib.setString(data[i], "%s", str)

Such an API is likely to be implemented in the near future.

Last, a future commit needs to deal with a small memory leak on the
static vector of DatasetTypeInfo elements; we're initializing that
vector with a call to `H5Tcopy()` that needs to be disposed prior to
closing the library. Again, this will be taken care of in the near term.