Support multiple string datasets as input/output
The DatasetInfo class has been revamped so that its constructor takes a
copy of the given dataset type and the destructor calls H5Tclose() on it.
Such a little change requires most vector appends of DatasetInfo entries
to be handled by std::move() -- otherwise the destructor closes the
borrowed handle, causing calls to the original DatasetInfo handle to fail.

A helper method DatasetInfo::reopenDatatype() has been introduced to
perform a call to H5Topen() on the private handle; that is useful on
cases where std::move() cannot be used. We also have to pass DataseInfo
variables by reference (as opposed to value) to function calls to avoid
the destructor from being called when it should not, too. In fact, we
should have been using references a long time ago!