Tests: allow different output for h5dump depending on the backend
It helps to have tests outputting slightly different data depending on
the backend. This is especially useful to deal with languages that opt
for a 1-based index instead of a 0-based one.

This commit also renames the expected output files so that they don't
end on '.lua', '.cpp', or '.py' -- but rather on '-lua', '-cpp', or
'-py'.