Add support for outputting compound datatypes
HDF5-UDF is now able to generate compound datasets on-the-fly. For
instance, it is possible to retrieve data from a CSV file and format it
as an HDF5 compound dataset or read data from sensors and output them in
a format that suits best the final application.

Compound datasets are described using a variation of the original
command-line syntax:

    `name:{member:type[,member:type...]}:resolution`

And with an example:

    `SensorData:{timestamp:int64,value:float}:1000`

UDFs that generate compounds are quite simple to write, as shown
by the two examples included in this commit.