Introduce strings with user-configurable size
Prior to this change, strings output by UDFs were fixed at 32
characters. Now, it is possible to specify with a '(N)' modifier how
many characters one wishes to reserve for such datatypes. The new
feature is supported by both string datasets and for compounds that
include a string member.

This code change also introduces support for a helper `lib.setString`
API that takes into account how many bytes can be copied from the input
string to the output buffer.

Note that LuaJIT seems to misbehave when strings with a size other than
32 are used, at least based on the examples shipped with this project.
This needs further investigation as to where the problem is coming from.