Sandbox: allow reading from regular CSV files
The ability to convert CSV files to HDF5 on-the-fly was the primary
reason for introducing support for outputting compound data types. Since
we don't have a configuration file yet, this commit relaxes the
sandboxing rules by letting files ending on '.csv' to be read by UDFs.
The limitations we impose are that they cannot be symlinks to other
objects in the filesystem and that they must be placed in the same
directory as where the HDF5 file is found.

We're going to make such rules configurable in the next major release
of HDF5-UDF.