readh5: support more image-based datasets
readh5 has been instrumented so it could render frames from the Tux
dataset and from the Doom demo in colors and horizontally stretched.
Now, with the introduction of other demos (like plotting the current
readings from weather stations) it makes sense to make support for
palette-based image rendering more generic. We do so by simply checking
whether a palette was provided in the command line (as opposed to
checking the dataset name).

This commit also fixes some HDF5 descriptor leaks. Not a big deal, since
this is a short-lived program, but it's nice to quiet HDF5 warnings when
running it in debug mode.