Include key validation features (#13)
Add initial support for signing UDFs with private keys. Management of public and private keys is currently limited to a single machine: a user is able to sign an UDF and to read back that same UDF by validating authorship through public key checks. The keys are stored on `~/.config/hdf5-udf`.

Co-authored-by: Thiago-Silva-das-Merces <Thiago.Silva.das.Merces@ibm.com>