JSON: new option to omit dump of source_code field
With the introduction of a new metadata field that holds a copy of the
UDF in source code form, dumping the JSON can be quite verbose. This
commit introduces a maximum length field to json.hpp's `dump()` method
which we use in `main()` to prevent printing strings that exceed 40
characters. Such strings are replaced by the '<omitted>' word instead.

As a minor change, we rename the variable from 'sourcecode' to
'source_code' and update the declaration of the `DummyBackend` in
sandbox.cpp so that it uses the new `compile()` API.