Incorporate public_key into JSON metadata
When signing a UDF, the associated public key is now embedded into the
JSON as extra metadata. That metadata is then automatically extracted
once the dataset is read for the first time on a different machine and
imported into that machine's ~/.config/hdf5-udf/deny profile.

By using the 'deny' profile by default, UDFs signed by that key will
not be allowed to execute any syscalls (once support for custom seccomp
rules is merged). A new  tool will be implemented at a later point to
assist users on the management of keys and their profiles.

Refs #17