Introduce a config file
To date, HDF5-UDF had hardcoded rules stating which syscalls a UDF was
able to run. This commit introduces support for configuration files
(written in JSON) that can be associated with groups of public key
files. The layout for such configuration files is as follows:

```
$HOME/
  .config/
    hdf5-udf/
      default/
        default.json
        some_key.pub
      deny/
        deny.json
        another_key.pub
      allow/
        allow.json
        one_another_key.pub
        yet_another_key.pub
```

JSON files must be named after the directory name (i.e., a directory
named 'allow' is expected to have a config file named 'allow.json' so
that in the event of having backups of old versions HDF5-UDF knows
which to pick). The structure of the JSON file is:

```
{
   "#": "Comments are allowed by starting a key with the '#' character",

   # "sandbox" option:
   # Set to 'false' to allow all syscalls to execute.
   # Set to 'true' to restrict which syscalls can run by filling in
   # the "syscalls" array. If set to "true" and no "syscalls" element
   # exist, then no syscalls are allowed to execute at all.
   "sandbox": boolean,

   # By default, all system calls are blocked. Each syscall you'd like
   # to allow to execute must be entered on this array.
   "syscalls": [
       # Set to 'true' to allow the given system call name to execute
       # regardless of which arguments are provided to that function.
       {"syscall_name": boolean}

       # In some cases it's useful to allow a certain syscall to
       # execute, as long as only harmless arguments are provided by the
       # user (e.g., O_RDONLY may be OK, whereas O_RDWR may not). The
       # following syntax allows one to state which arguments are valid
       # when enabling that syscall. Some notes:
       # - "arg" is the syscall argument number, starting from 0
       # - "op" is the operation, currently limited to "equals" and
       #   "is_set". The latter is useful to check if a given flag is
       #   set on a bitmask.
       # - "value" is the actual value to be tested. Strings are used to
       #   enter mnemonics such as TCGETS and O_TRUNC. Path names are not
       #   supported due to restrictions in the Seccomp Linux framework.
       {"syscall_name": {"arg": number, "op": string, "value": string|number}}
   ]
}
```

When **creating** a new UDF for the first time, HDF5-UDF creates the
directory structure under $HOME/.config/hdf5-udf and then generates a
new private and public keys for the currently logged in user.

When a signed UDF is read for the first time and its public key has not
been imported into the system before, HDF5-UDF creates the configuration
directory structure and automatically imports the public key into the
"deny" profile. This guarantees that, without prior user consent, no
foreign UDFs will ever be able to execute syscalls.

If HDF5-UDF detects that the UDF's public key matches that of the owner
of the private and public keys under $HOME/.config/hdf5-udf, then it
places no restrictions on which syscalls that UDF can execute. If needed,
the owner can be placed under a restricted list by moving the public key
file to the desired subdirectory (e.g., "allow", "deny", "default") that
contains the rules to apply.

For now, the code is able to identify the configuration file associated
with a given public key and to parse it. We are not yet translating the
parsed rules into seccomp calls -- that's left for the next patchset.

Last, the code changes in this patchset use the new <filesystem> header
available on C++17 onwards -- so we now build with -std=c++17. Old
compilers won't be able to build this code base.

Refs #17