Json: modify pretty-print for array elements
Objects embedded on arrays are always pretty-printed with a curly brace
followed by a newline (even when there's a single element to be
printed). This commit changes that behavior so that regular printing is
used when dealing with objects in arrays. This code change will become
handy when it comes the time to dynamically generate config files from
HDF5-UDF itself.

Refs #17