Dynamically generate configuration files
Configuration files for the three default profiles ("allow", "deny",
and "default") are now programmatically created along with the config
directory tree. This removes the need to ship default configuration
files and to incorporate logic to account for package relocation.

Prior to this change, SignatureHandler::validateProfileRules() also
played a role of patching the original JSON. However, as it turns
out, that's not as easy as it sounds: the modifications affect the
serialized data only. To make things easier, we're just checking
the data type when consuming elements at sandbox_init_seccomp()
and acting differently depending on whether we have a string or
a number on our hands.

This commit also reverts the introduction of a dependency on c++17,
as that adds more burden on maintenance and testing due to limited
availability of modern toolchain packages on certain distros. We're
now reimplementing the convenience bits of <filesystem> by ourselves.

Refs #17