Improve option parsing, make source_code optional
Storing the source code may add up a bit to the final metadata size,
so we make it optional (--append-sourcecode). We've also increased
the maximum JSON object size to be printed so that the public key
hash is always printed on the command line. Last, flags are now
recognized regardless of where in the command line they're entered.