hdf5-udf: reshape into a library form
Move the bulk of `main.cpp` into a set of high-level C-like APIs. This
is the first step towards having a shared library that third-party
applications can use. As a next step, `meson.build` needs to be modified
to produce and install such a shared library.

Note that the API is not set on stone yet, but serves as a starting
point so we can figure which features need to be exposed and which ones
can go away.

Refs #12, #15, and #17.