Python: handle nested sessions
Now that we have an API it's possible to call routines from HDF5-UDF
from a Python interpreter (such as a Jupyter notebook). In that case,
the caller will already have an active Python session, so we don't want
to execute Py_Finalize() and collapse the caller.

The right way to handle such situations is to create a new thread state
object and use that to execute the UDF -- which we do now.

Refs #20