Template: embed template strings in the library
Instead of shipping templates as text files that have to be searched for
at runtime, this commit incorporates them into libhdf5-udf.so. This
improves our ability to relocate the package, as we no longer rely on
searching for templates under $exec/../share/hdf5-udf. Also, dynamically
searching for those files is fragile, as we now have a library that can
be linked to other executables that live in a different directory than
`hdf5-udf`.

Refs #20