Python bindings: first commit
This is the first version of Python bindings for HDF5-UDF. We provide a
module named `hdf5_udf` which implements a `UserDefinedFunction` class.
The constructor currently expects both the path to the HDF5 and to the
UDF file. Methods allow one to specify extra options, to describe the
dataset, to compile the UDF, and to store the resulting bytecode to the
HDF5 file.

The API is still subject to change. In special, I'd like to see either
`compile()` or `store()` returning the resulting JSON. It may also make
sense to merge those two functions. That's to be reconsidered given the
use cases we have in mind.

The example below shows how simple it is to define a UDF with the
current bindings:

```
import inspect
from hdf5_udf import UserDefinedFunction

def dynamic_dataset():
    data = lib.getData("simple")
    dims = lib.getDims("simple")
    for i in range(dims[0]):
        data[i] = i

with open("/tmp/udf.py", "w") as f:
    f.write(inspect.getsource(dynamic_dataset))

with UserDefinedFunction(hdf5_file="/tmp/empty.h5", udf_file="/tmp/udf.py") as udf:
    udf.push_dataset({"name": "simple", "datatype": "float", "resolution": [2000]})
    udf.compile()
    udf.store()
```

Refs #20