Python: we have a package!
Our previous Python module extension was built in such a way that it
produced a shared library which was linked against libhdf5-udf.so. The
problem of having a shared library is that it restricts the list of
supported operating systems (and their versions) that the package can be
installed. After all, we have a binary package that expects the system
libraries to present certain symbols at runtime, and it's likely that
those requirements won't be met at all times.

There is a standard build system for shipping binary packages like this
called "manylinux". It is basically a Docker container that runs an
ancient version of CentOS; packages build on that environment are
guaranteed to be supported on more recent versions of that operating
system (and other mainline distros). Still, one cannot link to arbitrary
libraries such as libhdf5-udf.so: there is a small list of libraries
that manylinux can handle.

The approach we follow is to simply generate the FFI stubs needed to
call `dlopen(libhdf5-udf.so)` by ourselves. This way we can ship Python
bindings that are distro- and architecture-agnostic -- and we can
finally land on pypi.org!