API: new function to retrieve output and error messages
Move away from using fprintf(stderr) to storing error messages in the
UDF context combined with the creation of a new `libudf_get_error()`
function.

We also introduce a modification to `libudf_store()` so that the user
can retrieve the JSON metadata stored on the HDF5 file.

The Python bindings and the docs have been adapted accordingly.