API: libudf_store() changes
Let the function receive a pointer to a size_t that holds, on input, the
number of bytes allowed to be written to 'metadata' and, on output, the
actual number of bytes written to it.

The output of that function is now an array of JSON objects, too, to
account for UDFs that export two or more datasets.

The main utility and the Python bindings have been changed to adjust to
the new API. Moreover, `main` now retrieves and prints error messages
using `libudf_get_error()`.

We also remove some debug traces from the original code base.