API: introduce libudf_get_metadata()
This new API retrieves the JSON metadata associated with an UDF dataset.
Due to limitations in the I/O filter API, we resort to environment
variables to instruct the filter about the semantic behavior of the read
path: if a variable named `IOFILTER_READ_METADATA` is set, the only the
metadata returns to the caller. Otherwise, the UDF executes and the
output of that execution is provided to the caller.

With this API in place and the corresponding Python binding implemented,
we have all we need to retrieve the original source code from the HDF5
file.

Closes #15