Config dir: honor $XDG_CONFIG_HOME if it's set
The config dir defaults to `$XDG_CONFIG_HOME/hdf5-udf`, or
`$HOME/.config/hdf5-udf` if the `$XDG_CONFIG_HOME` is not set. If
neither variables exist, then `/tmp/hdf5-udf.username` is used instead.
A fallback `/tmp/hdf5-udf` is used if the username could not be
retrieved (hardly a problem on a non-networked user authentication
system)