Signature: improve public key search logic
We first search for configdir/*.pub + configdir/*/*.pub. If no matches
are found, we import the given key (associating it to the 'deny'
profile) and recurse to search again.

If at least one key has been found, we test each public key against the
one provided to the function. If there is a match, then the profile
associated with that key is returned. Otherwise, we import that key
(associating it to the 'deny' profile) and recurse to try again.