Signature: use JSON for files
Merge public and metadata files into one, while using JSON as
representation (as opposed to having the key as a binary file along with
a JSON representation of the metadata). The private key is also saved as
a JSON file now.

We remove a few verbose messages as part of this commit, too, and remove
the explicit 'deny' syscalls from the 'deny' profile so it's shorter and
easier to read.