Sandbox: improvements and fixes to config file parser
The "is_set" macro is not a good enough abstraction to deal with
bitmaps, as some definitions such as O_RDONLY have their values defined
as 0 and cause trouble when checking if that given value is set on a mask.

This patch exposes the very same interface already used by Seccomp: we
now take a reference bitmap mask ("mask") along with the bitmap value
to check on that mask. The original opcode has been renamed from
"is_set" to "masked_equals" as part of this modification.

This patch also fixes a silly bug with the identification of the opcode:
string.compare() returns 0 on a match.

Last, we improve the signal handler configuration by removing SIGSYS
from the list of blocked signals before installing the rules.