Sandbox: control filesystem access with ptrace, enhance config file
This commit adds support for monitoring the UDF process with Ptrace. The
parent process attaches to the child and then monitors the system calls
it attempts to execute. System calls that take a filesystem path as an
argument (such as `open()` and `rename()`) have their arguments
inspected by the parent. If the path has not been allowed to be accessed
by that UDF, or if the UDF attempted to open a file in write-mode while
read-only mode has been granted, then a SIGKILL is sent to the child.

We also extend the original per-profile config file by allowing one to
describe the filesystem paths an UDF is allowed to access. Basic regex
support is provided (including support for recursive directories). Some
examples:

- Allow read-write access to the root filesystem
```
{
    "filesystem": [
        {"/**": "rw"}
    ]
}
```

- Allow read-only access to the root filesystem
```
{
    "filesystem": [
        {"/**": "ro"}
    ]
}
```

- Allow acess to Python modules and basic system libraries
```
{
    "filesystem": [
        {"/**/python*/site-packages/**": "ro"},
        {"/**/libm*": "ro"},
        {"/**/libdl*": "ro"},
        {"/**/ld-linux*": "ro"},
        {"/**/libssl*": "ro"},
        {"/**/libcrypto*": "ro"},
        {"/**/libpthread*": "ro"}
    ]
}
```

While there are some known performance penalties with Ptrace (and even
ways to trick it with *time-to-check to time-of-use* race conditions),
Ptrace is a good candidate for replacing Syscall-Intercept. In special,
it doesn't require superuser capabilities to use it -- as opposed to
other solutions based on eBPF and AppArmor.

We may still want to have alternative backends at some point, but I'm
happy with this approach for now.

Closes #21